refactor(handlers): move shared post-save/discovery helpers to leaf modules (Stage B prep)
inactiveSyntaxDiagnostic + tryPostSaveSyntaxCheck + DDIC_POST_SAVE_CHECK_TYPES
→ write-helpers.ts; isTablesEndpointAvailable → feature-cache.ts. These are
shared by the read/write/activate handlers; relocating them to leaf modules
unblocks clean per-handler extraction (no handler→intent cycles). 3672 tests,
A1 zero-diff.
